Consolidate array Dispatch objects (#7505)
